+++ casanovo/config.py
@@ -102,6 +102,7 @@ class Config:
         replace_isoleucine_with_leucine=bool,
         massivekb_tokenizer=bool,
         residues=dict,
+        new_token_init=dict,
     )
 
     def __init__(self, config_file: Optional[str] = None):